changing TransportInfrastructure function properties into methods
