+++ src/NServiceBus.AcceptanceTests/FakeTransport/FakeTransport.cs
@@ -12,19 +12,17 @@
     {
         protected override TransportInfrastructure Initialize(SettingsHolder settings)
         {
-            return new FakeTransportInfrastructure(
-                Enumerable.Empty<Type>(),
-                TransportTransactionMode.ReceiveOnly,
-                new OutboundRoutingPolicy(OutboundRoutingType.Unicast, OutboundRoutingType.Unicast, OutboundRoutingType.Unicast),
-                s => new TransportSendInfrastructure(() => new FakeDispatcher(), () => Task.FromResult(StartupCheckResult.Success)),
-                s => new TransportReceiveInfrastructure(() => new FakeReceiver(settings.Get<Exception>()), () => new FakeQueueCreator(), () => Task.FromResult(StartupCheckResult.Success)));
+            return new FakeTransportInfrastructure(settings);
         }
     }
 
     public class FakeTransportInfrastructure : TransportInfrastructure
     {
-        public FakeTransportInfrastructure(IEnumerable<Type> deliveryConstraints, TransportTransactionMode transactionMode, OutboundRoutingPolicy outboundRoutingPolicy, Func<string, TransportSendInfrastructure> configureSendInfrastructure, Func<string, TransportReceiveInfrastructure> configureReceiveInfrastructure = null, Func<TransportSubscriptionInfrastructure> configureSubscriptionInfrastructure = null) : base(deliveryConstraints, transactionMode, outboundRoutingPolicy, configureSendInfrastructure, configureReceiveInfrastructure, configureSubscriptionInfrastructure)
+        readonly ReadOnlySettings settings;
+
+        public FakeTransportInfrastructure(ReadOnlySettings settings)
         {
+            this.settings = settings;
         }
 
         public override EndpointInstance BindToLocalEndpoint(EndpointInstance instance, ReadOnlySettings settings)
@@ -37,6 +35,25 @@
             return logicalAddress.ToString();
         }
         public override bool RequiresConnectionString => false;
+
+        public override TransportReceiveInfrastructure ConfigureReceiveInfrastructure(string connectionString)
+        {
+            return new TransportReceiveInfrastructure(() => new FakeReceiver(settings.Get<Exception>()), () => new FakeQueueCreator(), () => Task.FromResult(StartupCheckResult.Success));
+        }
+
+        public override TransportSendInfrastructure ConfigureSendInfrastructure(string connectionString)
+        {
+            return new TransportSendInfrastructure(() => new FakeDispatcher(), () => Task.FromResult(StartupCheckResult.Success));
+        }
+
+        public override TransportSubscriptionInfrastructure ConfigureSubscriptionInfrastructure()
+        {
+            throw new NotImplementedException();
+        }
+
+        public override IEnumerable<Type> DeliveryConstraints { get; } = Enumerable.Empty<Type>();
+        public override TransportTransactionMode TransactionMode { get; } = TransportTransactionMode.ReceiveOnly;
+        public override OutboundRoutingPolicy OutboundRoutingPolicy { get; } = new OutboundRoutingPolicy(OutboundRoutingType.Unicast, OutboundRoutingType.Unicast, OutboundRoutingType.Unicast);
         public override string ExampleConnectionStringForErrorMessage => null;
     }
 }
