+++ src/NServiceBus.Core/Transports/Msmq/MsmqTransportInfrastructure.cs
@@ -3,6 +3,10 @@ namespace NServiceBus
     using System;
     using System.Collections.Generic;
     using System.Text;
+    using System.Threading.Tasks;
+    using System.Transactions;
+    using NServiceBus.Features;
+    using NServiceBus.Performance.TimeToBeReceived;
     using NServiceBus.Routing;
     using NServiceBus.Settings;
     using NServiceBus.Transports;
@@ -12,18 +16,26 @@ namespace NServiceBus
     /// </summary>
     public class MsmqTransportInfrastructure : TransportInfrastructure
     {
-        /// <summary>
-        /// MsmqTransportInfrastructure.
-        /// </summary>
-        /// <param name="deliveryConstraints">The delivery constraints.</param>
-        /// <param name="transactionMode">The transaction mode.</param>
-        /// <param name="outboundRoutingPolicy">The outbound routing policy.</param>
-        /// <param name="configureSendInfrastructure">The send infrastructure factory.</param>
-        /// <param name="configureReceiveInfrastructure">The receive infrastructure factory.</param>
-        /// <param name="configureSubscriptionInfrastructure">The subscription infrastructure factory.</param>
-        public MsmqTransportInfrastructure(IEnumerable<Type> deliveryConstraints, TransportTransactionMode transactionMode, OutboundRoutingPolicy outboundRoutingPolicy, Func<string, TransportSendInfrastructure> configureSendInfrastructure, Func<string, TransportReceiveInfrastructure> configureReceiveInfrastructure = null, Func<TransportSubscriptionInfrastructure> configureSubscriptionInfrastructure = null) : base(deliveryConstraints, transactionMode, outboundRoutingPolicy, configureSendInfrastructure, configureReceiveInfrastructure, configureSubscriptionInfrastructure)
+        ReceiveStrategy SelectReceiveStrategy(TransportTransactionMode minimumConsistencyGuarantee, TransactionOptions transactionOptions)
+        {
+            if (minimumConsistencyGuarantee == TransportTransactionMode.TransactionScope)
+            {
+                return new ReceiveWithTransactionScope(transactionOptions);
+            }
+
+            if (minimumConsistencyGuarantee == TransportTransactionMode.None)
+            {
+                return new ReceiveWithNoTransaction();
+            }
+
+            return new ReceiveWithNativeTransaction();
+        }
+
+        internal MsmqTransportInfrastructure(ReadOnlySettings settings)
         {
             RequireOutboxConsent = true;
+
+            this.settings = settings;
         }
 
         /// <summary>
@@ -32,7 +44,7 @@ namespace NServiceBus
         public override EndpointInstance BindToLocalEndpoint(EndpointInstance instance, ReadOnlySettings settings) => instance.AtMachine(Environment.MachineName);
 
         /// <summary>
-        /// Converts a given logical address to the transport address.
+        /// <see cref="TransportInfrastructure.ToTransportAddress"/>.
         /// </summary>
         /// <param name="logicalAddress">The logical address.</param>
         /// <returns>The transport address.</returns>
@@ -57,7 +69,7 @@ namespace NServiceBus
         }
 
         /// <summary>
-        /// Returns the canonical for of the given transport address so various transport addresses can be effectively compared and deduplicated.
+        /// <see cref="TransportInfrastructure.MakeCanonicalForm"/>.
         /// </summary>
         /// <param name="transportAddress">A transport address.</param>
         public override string MakeCanonicalForm(string transportAddress)
@@ -66,13 +78,104 @@ namespace NServiceBus
         }
 
         /// <summary>
-        /// Gets an example connection string to use when reporting lack of configured connection string to the user.
+        /// <see cref="TransportInfrastructure.ConfigureReceiveInfrastructure"/>.
+        /// </summary>
+        /// <param name="connectionString">Connection string.</param>
+        /// <returns>Transport receive infrastructure.</returns>
+        public override TransportReceiveInfrastructure ConfigureReceiveInfrastructure(string connectionString)
+        {
+            new CheckMachineNameForComplianceWithDtcLimitation().Check();
+
+            var builder = connectionString != null
+                ? new MsmqConnectionStringBuilder(connectionString).RetrieveSettings()
+                : new MsmqSettings();
+
+            MsmqScopeOptions scopeOptions;
+
+            if (!settings.TryGet(out scopeOptions))
+            {
+                scopeOptions = new MsmqScopeOptions();
+            }
+
+            return new TransportReceiveInfrastructure(
+                () => new MessagePump(guarantee => SelectReceiveStrategy(guarantee, scopeOptions.TransactionOptions)),
+                () => new QueueCreator(builder),
+                () =>
+                {
+                    var bindings = settings.Get<QueueBindings>();
+                    new QueuePermissionChecker().CheckQueuePermissions(bindings.ReceivingAddresses);
+                    return Task.FromResult(StartupCheckResult.Success);
+                });
+        }
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.ConfigureSendInfrastructure"/>.
+        /// </summary>
+        /// <param name="connectionString">Connection string.</param>
+        /// <returns>Transport send infrastructure.</returns>
+        public override TransportSendInfrastructure ConfigureSendInfrastructure(string connectionString)
+        {
+            new CheckMachineNameForComplianceWithDtcLimitation().Check();
+
+            Func<IReadOnlyDictionary<string, string>, string> getMessageLabel;
+            settings.TryGet("Msmq.GetMessageLabel", out getMessageLabel);
+
+            Func<IReadOnlyDictionary<string, string>, string> messageLabelGenerator;
+            if (!settings.TryGet("msmqLabelGenerator", out messageLabelGenerator))
+            {
+                messageLabelGenerator = headers => string.Empty;
+            }
+
+            var builder = new MsmqConnectionStringBuilder(connectionString).RetrieveSettings();
+
+            return new TransportSendInfrastructure(
+                () => new MsmqMessageDispatcher(builder, messageLabelGenerator),
+                () =>
+                {
+                    var bindings = settings.Get<QueueBindings>();
+                    new QueuePermissionChecker().CheckQueuePermissions(bindings.SendingAddresses);
+                    var result = new MsmqTimeToBeReceivedOverrideCheck(settings).CheckTimeToBeReceivedOverrides();
+                    return Task.FromResult(result);
+                });
+        }
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.ConfigureSubscriptionInfrastructure"/>.
+        /// </summary>
+        /// <returns>Transport subscription infrastructure.</returns>
+        public override TransportSubscriptionInfrastructure ConfigureSubscriptionInfrastructure()
+        {
+            throw new NotImplementedException("MSMQ does not support native pub/sub.");
+        }
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.DeliveryConstraints"/>.
+        /// </summary>
+        public override IEnumerable<Type> DeliveryConstraints { get; } = new[]
+        {
+            typeof(DiscardIfNotReceivedBefore)
+        };
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.TransactionMode"/>.
+        /// </summary>
+        public override TransportTransactionMode TransactionMode { get; } = TransportTransactionMode.TransactionScope;
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.OutboundRoutingPolicy"/>.
+        /// </summary>
+        public override OutboundRoutingPolicy OutboundRoutingPolicy { get; } = new OutboundRoutingPolicy(OutboundRoutingType.Unicast, OutboundRoutingType.Unicast, OutboundRoutingType.Unicast);
+
+        /// <summary>
+        /// <see cref="TransportInfrastructure.ExampleConnectionStringForErrorMessage"/>.
         /// </summary>
         public override string ExampleConnectionStringForErrorMessage => "cacheSendConnection=true;journal=false;deadLetter=true";
 
         /// <summary>
-        /// Used by implementations to control if a connection string is necessary.
+        /// <see cref="TransportInfrastructure.RequiresConnectionString"/>.
         /// </summary>
         public override bool RequiresConnectionString => false;
+
+        readonly ReadOnlySettings settings;
     }
 }
